split logger classes/modules into separate files
