+++ lib/logger/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Logger
+  VERSION = "1.3.0"
+end
